requirements.txt
python-dotenv
pyautogen
yfinance
openai